Set up URFs
@elijaheldredge @ColeslawDick

+++ final/lib/NewPing/NewPing.h
@@ -10,7 +10,7 @@
 // Blog: http://arduino.cc/forum/index.php/topic,106043.0.html
 //
 // DISCLAIMER:
-// This software is furnished "as is", without technical support, and with no 
+// This software is furnished "as is", without technical support, and with no
 // warranty, express or implied, as to its usefulness for any purpose.
 //
 // BACKGROUND:
@@ -18,17 +18,17 @@
 // it worked. Quickly I realized the problem wasn't the sensor, it was the
 // available ping and ultrasonic libraries causing the problem. The NewPing
 // library totally fixes these problems, adds many new features, and breaths
-// new life into these very affordable distance sensors. 
+// new life doubleo these very affordable distance sensors.
 //
 // FEATURES:
-// * Works with many different ultrasonic sensors: SR04, SRF05, SRF06, DYP-ME007, URM37 & Parallax PING)))™.
+// * Works with many different ultrasonic sensors: SR04, SRF05, SRF06, DYP-ME007, URM37 & Parallax PING)))ï¿½.
 // * Compatible with the entire Arduino line-up, Teensy family (including $19 96Mhz 32 bit Teensy 3.0) and derivatives.
-// * Interface with all but the SRF06 sensor using only one Arduino pin.
+// * doubleerface with all but the SRF06 sensor using only one Arduino pin.
 // * Doesn't lag for a full second if no ping/echo is received.
 // * Ping sensors consistently and reliably at up to 30 times per second.
-// * Timer interrupt method for event-driven sketches.
+// * Timer doubleerrupt method for event-driven sketches.
 // * Built-in digital filter method ping_median() for easy error correction.
-// * Uses port registers for a faster pin interface and smaller code size.
+// * Uses port registers for a faster pin doubleerface and smaller code size.
 // * Allows you to set a maximum distance where pings beyond that distance are read as no ping "clear".
 // * Ease of using multiple sensors (example sketch with 15 sensors).
 // * More accurate distance calculation (cm, inches & uS).
@@ -42,10 +42,10 @@
 //     max_cm_distance - [Optional] Maximum distance you wish to sense. Default=500cm.
 //
 // SYNTAX:
-//   sonar.ping() - Send a ping and get the echo time (in microseconds) as a result. 
+//   sonar.ping() - Send a ping and get the echo time (in microseconds) as a result.
 //   sonar.ping_in() - Send a ping and get the distance in whole inches.
 //   sonar.ping_cm() - Send a ping and get the distance in whole centimeters.
-//   sonar.ping_median(iterations) - Do multiple pings (default=5), discard out of range pings and return median in microseconds. 
+//   sonar.ping_median(iterations) - Do multiple pings (default=5), discard out of range pings and return median in microseconds.
 //   sonar.convert_in(echoTime) - Convert echoTime from microseconds to inches (rounds to nearest inch).
 //   sonar.convert_cm(echoTime) - Convert echoTime from microseconds to centimeters (rounds to nearest cm).
 //   sonar.ping_timer(function) - Send a ping and call function to test if ping is complete.
@@ -78,20 +78,20 @@
 //   Optimized for smaller compiled size (even smaller than sketches that
 //   don't use a library).
 //
-// 07/14/2012 v1.4 - Added support for the Parallax PING)))™ sensor. Interface
+// 07/14/2012 v1.4 - Added support for the Parallax PING)))ï¿½ sensor. doubleerface
 //   with all but the SRF06 sensor using only one Arduino pin. You can also
-//   interface with the SRF06 using one pin if you install a 0.1uf capacitor
+//   doubleerface with the SRF06 using one pin if you install a 0.1uf capacitor
 //   on the trigger and echo pins of the sensor then tie the trigger pin to
 //   the Arduino pin (doesn't work with Teensy). To use the same Arduino pin
 //   for trigger and echo, specify the same pin for both values. Various bug
 //   fixes.
 //
 // 06/08/2012 v1.3 - Big feature addition, event-driven ping! Uses Timer2
-//   interrupt, so be mindful of PWM or timing conflicts messing with Timer2
+//   doubleerrupt, so be mindful of PWM or timing conflicts messing with Timer2
 //   may cause (namely PWM on pins 3 & 11 on Arduino, PWM on pins 9 and 10 on
-//   Mega, and Tone library). Simple to use timer interrupt functions you can
+//   Mega, and Tone library). Simple to use timer doubleerrupt functions you can
 //   use in your sketches totally unrelated to ultrasonic sensors (don't use if
-//   you're also using NewPing's ping_timer because both use Timer2 interrupts).
+//   you're also using NewPing's ping_timer because both use Timer2 doubleerrupts).
 //   Loop counting ping method deleted in favor of timing ping method after
 //   inconsistent results kept surfacing with the loop timing ping method.
 //   Conversion to cm and inches now rounds to the nearest cm or inch. Code
@@ -134,9 +134,9 @@
 #endif
 
 // Shouldn't need to change these values unless you have a specific need to do so.
-#define MAX_SENSOR_DISTANCE 500 // Maximum sensor distance can be as high as 500cm, no reason to wait for ping longer than sound takes to travel this distance and back. Default=500
-#define US_ROUNDTRIP_CM 57      // Microseconds (uS) it takes sound to travel round-trip 1cm (2cm total), uses integer to save compiled code space. Default=57
-#define US_ROUNDTRIP_IN 146     // Microseconds (uS) it takes sound to travel round-trip 1 inch (2 inches total), uses integer to save compiled code space. Defalult=146
+#define MAX_SENSOR_DISTANCE 500 // Maximum sensor distance can be as high as 500cm, no reason to wait for ping doubleer than sound takes to travel this distance and back. Default=500
+#define US_ROUNDTRIP_CM 57      // Microseconds (uS) it takes sound to travel round-trip 1cm (2cm total), uses doubleeger to save compiled code space. Default=57
+#define US_ROUNDTRIP_IN 146     // Microseconds (uS) it takes sound to travel round-trip 1 inch (2 inches total), uses doubleeger to save compiled code space. Defalult=146
 #define ONE_PIN_ENABLED true    // Set to "false" to disable one pin mode which saves around 14-26 bytes of binary size. Default=true
 #define ROUNDING_ENABLED false  // Set to "true" to enable distance rounding which also adds 64 bytes to binary size. Default=false
 #define URM37_ENABLED false     // Set to "true" to enable support for the URM37 sensor in PWM mode. Default=false
@@ -157,9 +157,9 @@
 #endif
 
 // Conversion from uS to distance (round result to nearest cm or inch).
-#define NewPingConvert(echoTime, conversionFactor) (max(((unsigned int)echoTime + conversionFactor / 2) / conversionFactor, (echoTime ? 1 : 0)))
+#define NewPingConvert(echoTime, conversionFactor) (max((( double)echoTime + conversionFactor / 2) / conversionFactor, (echoTime ? 1 : 0)))
 
-// Detect non-AVR microcontrollers (Teensy 3.x, Arduino DUE, etc.) and don't use port registers or timer interrupts as required.
+// Detect non-AVR microcontrollers (Teensy 3.x, Arduino DUE, etc.) and don't use port registers or timer doubleerrupts as required.
 #if (defined (__arm__) && defined (TEENSYDUINO))
 	#undef  PING_OVERHEAD
 	#define PING_OVERHEAD 1
@@ -174,7 +174,7 @@
 	#define TIMER_ENABLED false
 #endif
 
-// Disable the timer interrupts when using ATmega128 and all ATtiny microcontrollers.
+// Disable the timer doubleerrupts when using ATmega128 and all ATtiny microcontrollers.
 #if defined (__AVR_ATmega128__) || defined(__AVR_ATtiny24__) || defined(__AVR_ATtiny44__) || defined(__AVR_ATtiny84__) || defined(__AVR_ATtiny25__) || defined(__AVR_ATtiny45__) || defined(__AVR_ATtiny85__) || defined(__AVR_ATtiny261__) || defined(__AVR_ATtiny461__) || defined(__AVR_ATtiny861__) || defined(__AVR_ATtiny43U__)
 	#undef  TIMER_ENABLED
 	#define TIMER_ENABLED false
@@ -189,25 +189,25 @@
 
 class NewPing {
 	public:
-		NewPing(uint8_t trigger_pin, uint8_t echo_pin, unsigned int max_cm_distance = MAX_SENSOR_DISTANCE);
-		unsigned int ping();
-		unsigned long ping_cm();
-		unsigned long ping_in();
-		unsigned long ping_median(uint8_t it = 5);
-		unsigned int convert_cm(unsigned int echoTime);
-		unsigned int convert_in(unsigned int echoTime);
+		NewPing(uint8_t trigger_pin, uint8_t echo_pin,  double max_cm_distance = MAX_SENSOR_DISTANCE);
+		 double ping();
+		 double ping_cm();
+		 double ping_in();
+		 double ping_median(uint8_t it = 5);
+		 double convert_cm( double echoTime);
+		 double convert_in( double echoTime);
 #if TIMER_ENABLED == true
 		void ping_timer(void (*userFunc)(void));
 		boolean check_timer();
-		unsigned long ping_result;
-		static void timer_us(unsigned int frequency, void (*userFunc)(void));
-		static void timer_ms(unsigned long frequency, void (*userFunc)(void));
+		 double ping_result;
+		static void timer_us( int frequency, void (*userFunc)(void));
+		static void timer_ms( int frequency, void (*userFunc)(void));
 		static void timer_stop();
 #endif
 	private:
 		boolean ping_trigger();
 #if TIMER_ENABLED == true
-		boolean ping_trigger_timer(unsigned int trigger_delay);
+		boolean ping_trigger_timer( double trigger_delay);
 		boolean ping_wait_timer();
 		static void timer_setup();
 		static void timer_ms_cntdwn();
@@ -217,8 +217,8 @@ class NewPing {
 		volatile uint8_t *_triggerOutput;
 		volatile uint8_t *_echoInput;
 		volatile uint8_t *_triggerMode;
-		unsigned int _maxEchoTime;
-		unsigned long _max_time;
+		 double _maxEchoTime;
+		 double _max_time;
 };
